feat: load react email serverside
